backend/requirements.txt
gunicorn
pyjwt
aiohttp
aiodns
aiomysql
